Adds styling to buttons and reorganizes components vs containers
