+++ src/containers/MovieArea/MovieArea.scss
