+++ src/containers/MovieArea/MovieArea.test.js
