+++ src/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./components/App/App";
+import App from "./containers/App/App";
 import * as serviceWorker from "./serviceWorker";
 import rootReducer from "./reducers";
 import { createStore } from "redux";
