+++ src/index.scss
@@ -1,10 +1,10 @@
 @import url('https://fonts.googleapis.com/css?family=Roboto');
 @import '/variables';
 @import "./components/MovieCard/MovieCard";
-@import './components/CreateUser/CreateUser';
+@import './containers/CreateUser/CreateUser';
 @import './components/Header/Header';
-@import './components/MovieArea/MovieArea';
-@import './components/Login/Login';
+@import './containers/MovieArea/MovieArea';
+@import './containers/Login/Login';
 @import './components/MovieDetails/MovieDetails';
